+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 - Hide files starting with `.` in the timeline
 - Prevent automatically retrying failed indexing jobs
+- MySQL fixes: run `occ places-setup` again if you use MySQL (not MariaDB)
 
 ## [v6.2.2] - 2024-01-10